Show the uploaded front photo on certification wallet cards
