+++ lib/features/certifications/presentation/widgets/certification_ecard.dart
@@ -1,13 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
-import 'package:submersion/l10n/l10n_extension.dart';
 import 'package:submersion/features/certifications/domain/entities/certification.dart';
 import 'package:submersion/features/certifications/presentation/widgets/certification_ecard_back.dart';
+import 'package:submersion/features/certifications/presentation/widgets/certification_ecard_front.dart';
 
 /// A credit card-style widget displaying a certification with agency branding.
 ///
-/// Supports both front and back views with an animated flip transition.
+/// Supports both front and back views with an animated flip transition. Each
+/// face shows the diver's uploaded photo of that side when one exists, and a
+/// generated design otherwise.
 class CertificationEcard extends StatelessWidget {
   /// The certification to display.
   final Certification certification;
@@ -65,7 +67,7 @@ class CertificationEcard extends StatelessWidget {
                     key: const ValueKey('back'),
                     certification: certification,
                   )
-                : _CardFront(
+                : CertificationEcardFront(
                     key: const ValueKey('front'),
                     certification: certification,
                     diverName: diverName,
@@ -74,248 +76,5 @@ class CertificationEcard extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-/// The front face of the certification card.
-class _CardFront extends StatelessWidget {
-  final Certification certification;
-  final String diverName;
-
-  const _CardFront({
-    super.key,
-    required this.certification,
-    required this.diverName,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final agency = certification.agency;
-
-    return Container(
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(16),
-        gradient: LinearGradient(
-          begin: Alignment.topLeft,
-          end: Alignment.bottomRight,
-          colors: [agency.primaryColor, agency.secondaryColor],
-        ),
-        boxShadow: [
-          BoxShadow(
-            color: agency.primaryColor.withValues(alpha: 0.4),
-            blurRadius: 12,
-            offset: const Offset(0, 6),
-          ),
-        ],
-      ),
-      clipBehavior: Clip.antiAlias,
-      child: Stack(
-        children: [
-          // Decorative wave pattern
-          Positioned.fill(
-            child: CustomPaint(
-              painter: _WavePatternPainter(
-                color: Colors.white.withValues(alpha: 0.08),
-              ),
-            ),
-          ),
-          // Card content
-          Padding(
-            padding: const EdgeInsets.all(20),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                // Top row: agency name and status badge
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Expanded(
-                      child: Text(
-                        agency.displayName,
-                        style: const TextStyle(
-                          color: Colors.white,
-                          fontSize: 16,
-                          fontWeight: FontWeight.bold,
-                          letterSpacing: 1.2,
-                        ),
-                      ),
-                    ),
-                    _buildStatusBadge(context),
-                  ],
-                ),
-                const Spacer(),
-                // Center: certification name
-                Text(
-                  certification.name,
-                  style: const TextStyle(
-                    color: Colors.white,
-                    fontSize: 22,
-                    fontWeight: FontWeight.w600,
-                    letterSpacing: 0.5,
-                  ),
-                  maxLines: 2,
-                  overflow: TextOverflow.ellipsis,
-                ),
-                // Level display if present
-                if (certification.level != null) ...[
-                  const SizedBox(height: 4),
-                  Text(
-                    certification.level!.displayName,
-                    style: TextStyle(
-                      color: Colors.white.withValues(alpha: 0.85),
-                      fontSize: 14,
-                      fontWeight: FontWeight.w400,
-                    ),
-                  ),
-                ],
-                const Spacer(),
-                // Bottom row: diver info and issue date
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  crossAxisAlignment: CrossAxisAlignment.end,
-                  children: [
-                    Expanded(
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        mainAxisSize: MainAxisSize.min,
-                        children: [
-                          Text(
-                            diverName.toUpperCase(),
-                            style: const TextStyle(
-                              color: Colors.white,
-                              fontSize: 14,
-                              fontWeight: FontWeight.w600,
-                              letterSpacing: 1.5,
-                            ),
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                          if (certification.cardNumber != null &&
-                              certification.cardNumber!.isNotEmpty) ...[
-                            const SizedBox(height: 2),
-                            Text(
-                              certification.cardNumber!,
-                              style: TextStyle(
-                                color: Colors.white.withValues(alpha: 0.8),
-                                fontSize: 12,
-                                letterSpacing: 1.0,
-                              ),
-                              maxLines: 1,
-                              overflow: TextOverflow.ellipsis,
-                            ),
-                          ],
-                        ],
-                      ),
-                    ),
-                    const SizedBox(width: 16),
-                    if (certification.issueDate != null)
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.end,
-                        mainAxisSize: MainAxisSize.min,
-                        children: [
-                          Text(
-                            context.l10n.certifications_ecard_label_issued,
-                            style: TextStyle(
-                              color: Colors.white.withValues(alpha: 0.7),
-                              fontSize: 10,
-                              fontWeight: FontWeight.w500,
-                              letterSpacing: 1.0,
-                            ),
-                          ),
-                          const SizedBox(height: 2),
-                          Text(
-                            DateFormat(
-                              'MM/yy',
-                            ).format(certification.issueDate!),
-                            style: const TextStyle(
-                              color: Colors.white,
-                              fontSize: 14,
-                              fontWeight: FontWeight.w600,
-                            ),
-                          ),
-                        ],
-                      ),
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildStatusBadge(BuildContext context) {
-    if (certification.isExpired) {
-      return Container(
-        padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-        decoration: BoxDecoration(
-          color: Colors.red,
-          borderRadius: BorderRadius.circular(4),
-        ),
-        child: Text(
-          context.l10n.certifications_ecard_statusBadge_expired,
-          style: const TextStyle(
-            color: Colors.white,
-            fontSize: 10,
-            fontWeight: FontWeight.bold,
-            letterSpacing: 0.5,
-          ),
-        ),
-      );
-    }
-
-    if (certification.expiresWithin(90)) {
-      return Container(
-        padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-        decoration: BoxDecoration(
-          color: Colors.orange,
-          borderRadius: BorderRadius.circular(4),
-        ),
-        child: Text(
-          context.l10n.certifications_ecard_statusBadge_expiring,
-          style: const TextStyle(
-            color: Colors.white,
-            fontSize: 10,
-            fontWeight: FontWeight.bold,
-            letterSpacing: 0.5,
-          ),
-        ),
-      );
-    }
-
-    return const SizedBox.shrink();
-  }
-}
-
-/// Custom painter for decorative wave pattern on the card.
-class _WavePatternPainter extends CustomPainter {
-  final Color color;
-
-  _WavePatternPainter({required this.color});
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final paint = Paint()
-      ..color = color
-      ..style = PaintingStyle.fill;
-
-    // Draw decorative circles at various positions
-    final circles = [
-      (Offset(size.width * 0.85, size.height * 0.2), size.width * 0.25),
-      (Offset(size.width * 0.95, size.height * 0.6), size.width * 0.18),
-      (Offset(size.width * 0.1, size.height * 0.9), size.width * 0.15),
-      (Offset(size.width * 0.75, size.height * 0.85), size.width * 0.12),
-    ];
-
-    for (final (offset, radius) in circles) {
-      canvas.drawCircle(offset, radius, paint);
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant _WavePatternPainter oldDelegate) {
-    return oldDelegate.color != color;
   }
 }
